feat: Added token and ast file location for better error display

+++ src/lexer/token.rs
@@ -3,7 +3,7 @@
 #[derive(Debug, Eq, PartialEq, Clone)]
 pub struct Token {
     pub line: u32,
-    pub column: u16,
+    pub column: u32,
     pub token_type: TokenType,
 }
